[ui/#299] 채팅 메시지 목록 컴포넌트화
